Implement voice transcrption using mozilla/DeepSpeech.
I've implemented a basic form of voice transcription using mozilla/DeepSpeech pulling from the examples at mozilla/DeepSpeech-examples.

More work will be required in order to improve the accuracy of the model; it seems to fail frequently with my accent and is unable to pick up certain words at all (e.g. 'urology' or 'radiotherapy', only managing 'radio' and 'therapy' with a pronounced pause in between).