Added basic text-to-speech functionality.
I've added basic text-to-speech functionality that simply response what the user has said.

I've also fixed a bug where the buffer_queue wasn't getting cleared, so the voice assistant would listen to the output of the TTS. A welcome side-effect is that accuracy seems to have improved at the same time.

Currently, the config will use a British-voiced TTS speaker on Windows machines with a British English language pack. The configuration needs modification on other platforms for this functionality.

We may not stick with this library in the long run, but it seems useful in building an MVP!

I've also renamed some files to make more sense semantically.